+++ system/info_windows.go
@@ -74,6 +74,12 @@ func GetSystemInfo() (*SystemInfo, error) {
 	return nil, nil
 }
 
+// ❗ GetOSInfo returns info about OS
+func GetOSInfo() (*OSInfo, error) {
+	panic("UNSUPPORTED")
+	return nil, nil
+}
+
 // ❗ GetInterfacesStats returns info about network interfaces
 func GetInterfacesStats() (map[string]*InterfaceStats, error) {
 	panic("UNSUPPORTED")