+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `[tmp]` Use `os.TempDir` for default directory instead of hardcoded path
 * `[system]` Fixed stubs for Windows
 * `[terminal]` Fixed stubs for Windows
+* `[log]` Improved documentation
 
 ### 12.45.0